Integrating piga into the input system.

+++ include/SDLEventHandler.hpp
@@ -14,11 +14,11 @@ class SDLEventHandler
         typedef sigc::signal<void, const SDL_Event&, float> SDLEventSignal;
        	void sendEvent(const SDL_Event &e, float frametime);
 
-        SDLEventSignal getSignal(uint8_t type);
+        SDLEventSignal getSignal(int32_t type);
     protected:
         virtual void onSDLEvent(const SDL_Event &e, float frametime);
     private:
-		std::map<uint8_t, SDLEventSignal> m_signals;
+        std::map<uint32_t, SDLEventSignal> m_signals;
 };
 
 #endif
